:hammer: Extract ControlUtils methods so they can be used by common module
